renamed OnFundsTransferResponse to RateResponse and moved it from hooks.rs (which has been deleted) to rates.rs in the ado_base folder
