+++ contracts/finance/andromeda-weighted-distribution-splitter/src/mock_querier.rs
@@ -1,5 +1,5 @@
 use andromeda_app::app::QueryMsg as AppQueryMsg;
-use common::{ado_base::hooks::OnFundsTransferResponse, Funds};
+use common::{ado_base::hooks::RatesResponse, Funds};
 use cosmwasm_std::{
     from_json,
     testing::{mock_env, MockApi, MockQuerier, MockStorage, MOCK_CONTRACT_ADDR},
@@ -131,7 +131,7 @@ impl WasmMockQuerier {
                             ))
                         }
                     };
-                    let response = OnFundsTransferResponse {
+                    let response = RatesResponse {
                         msgs,
                         events: vec![],
                         leftover_funds: new_funds,
